+++ ProjectSMA/src/jade/ProductB.java
@@ -1,5 +1,10 @@
 package jade;
 
-public class ProductB extends Product { 
-    public ProductB(float price, float quantity) { super(price, quantity); } 
-} 
+public class ProductB extends Product {
+	
+	public ProductB(double price, int quantity) {
+		
+		super(price, quantity);
+		
+	}
+}
