+++ ProjectSMA/src/jade/ProductC.java
@@ -1,5 +1,10 @@
 package jade;
 
-public class ProductC extends Product { 
-    public ProductC(float price, float quantity) { super(price, quantity); } 
-} 
+public class ProductC extends Product {
+
+	public ProductC(double price, int quantity) {
+
+		super(price, quantity);
+
+	}
+}
